pipecat-agent/requirements.txt
pipecat-ai[webrtc,daily,silero,deepgram,openai,cartesia,runner]
python-dotenv
loguru
